requirements.txt
matplotlib==3.2.1
mrcfile==1.3.0
numpy==1.16.4
open3d==0.14.1
# open3d==0.11.2
opencv-python==4.4.0.44
pandas==1.1.1
Pillow==6.0.0
plotly==4.9.0
python-dateutil==2.8.1
pytorch-fid==0.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-video==1.1.11
scipy==1.2.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
zipp==3.1.0
torch-fidelity==0.2.0
easydict==1.9
tensorboardX==2.2
lmdb==1.1.1
sk-video==1.1.10
lpips==0.1.3
torch-ema==0.3.0
# torch-ema@git+https://github.com/fadel/pytorch_ema@3950a7b5c4b88f46fd14f620277bad21898597a9